+++ libguile/discouraged.h
@@ -90,11 +90,11 @@ SCM_API scm_t_ptrdiff scm_num2ptrdiff (SCM num, unsigned long int pos,
 SCM_API size_t scm_num2size (SCM num, unsigned long int pos,
 			     const char *s_caller);
 #if SCM_SIZEOF_LONG_LONG != 0
-SCM_API SCM scm_long_long2num (long long sl);
-SCM_API SCM scm_ulong_long2num (unsigned long long sl);
-SCM_API long long scm_num2long_long (SCM num, unsigned long int pos,
+SCM_API SCM scm_long_long2num (int64_t sl);
+SCM_API SCM scm_ulong_long2num (uint64_t sl);
+SCM_API int64_t scm_num2long_long (SCM num, unsigned long int pos,
 				     const char *s_caller);
-SCM_API unsigned long long scm_num2ulong_long (SCM num, unsigned long int pos,
+SCM_API uint64_t scm_num2ulong_long (SCM num, unsigned long int pos,
 					       const char *s_caller);
 #endif
